+++ package.json
@@ -7,6 +7,7 @@
     "node-sass": "^4.12.0",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
+    "react-firebase-hooks": "^2.1.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0"
   },
